move rect shader to rect renderer
